pyproject.toml
[project]
name = "bank-app"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "laksumi213", email = "laksumi.fb@gmail.com" }
]
dependencies = [
    "streamlit",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "selenium>=4.39.0",
    "webdriver-manager>=4.0.2",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "duckduckgo-search>=8.1.1",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "gspread-dataframe>=4.0.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/streamlit"]
